+++ src/SeerConfigDialog.h
@@ -44,6 +44,10 @@ class SeerConfigDialog : public QDialog, protected Ui::SeerConfigDialogForm {
         bool                                seerClearManualCommandHistory                   () const;
 
         // Gdb settings.
+        // Program to launch gdb, rr, etc. Can be blank. Needed by flatpak.
+        void                                setGdbLauncher                                  (const QString& launcher);
+        QString                             gdbLauncher                                     () const;
+
         void                                setGdbProgram                                   (const QString& program);
         QString                             gdbProgram                                      () const;