+++ src/SeerGdbConfigPage.h
@@ -17,6 +17,7 @@ class SeerGdbConfigPage : public QWidget, protected Ui::SeerGdbConfigPage {
        ~SeerGdbConfigPage ();
 
         QString                 gdbProgram                                      () const;
+        QString                 gdbLauncher                                     () const;
         QString                 gdbArguments                                    () const;
         bool                    gdbAsyncMode                                    () const;
         bool                    gdbNonStopMode                                  () const;
@@ -28,6 +29,7 @@ class SeerGdbConfigPage : public QWidget, protected Ui::SeerGdbConfigPage {
 
 
         void                    setGdbProgram                                   (const QString& program);
+        void                    setGdbLauncher                                  (const QString& launcher);
         void                    setGdbArguments                                 (const QString& arguments);
         void                    setGdbAsyncMode                                 (bool flag);
         void                    setGdbNonStopMode                               (bool flag);